backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
whitenoise==6.6.0
python-decouple==3.8
django-filter==23.5
Pillow==11.0.0